# Speed up the `dashboardapi/section_courses/:section_id` endpoint