Speeds up the `dashboardapi/section_courses/:section_id` endpoint by de-duplicating database queries and simplifying the code structure.

the section_courses endpoint previously did this:
Retrieve the list of unit_ids that students have made progress in
For ALL CourseOfferings: check if any of the CourseOfferings's CourseVersions are a unit and if that unit is in the list of unit_ids
For ALL CourseVersions: if that CourseVersion is a UnitGroup (not a Unit), check if any Units in that UnitGroup are in the list of unit_ids
Return the concatenation of those two lists

Now:
Retrieve the list of unit_ids that students have made progress in
For each unit in unit_ids, find all corresponding CourseOfferings that are Units through associations
for each unit in unit_ids, find all corresponding CourseVersions that are UnitGroups through associations
return the concatenation

This significantly decreases the number of database queries

Old queries:
1278 queries, example completed in `1589ms (Views: 7.4ms | ActiveRecord: 549.7ms | Allocations: 1450020)`
Another example with just units: 1333 queries
Another example with just unit groups: 300 queries

New queries: 168 queries, example completed in `174ms (Views: 3.5ms | ActiveRecord: 44.8ms | Allocations: 79084)`
Another example with just units: 15 queries
Another example with just unit groups: 71 queries

## Links
https://codedotorg.atlassian.net/browse/TEACH-1515
<!--
  Links to relevant external resources; ie, specification documents, Jira tickets, related PRs, Honeybadger errors, etc.
-->