@lfryemason 2025-01-21
@etaderhold we already have a test for a bunch of different kinds of units, including no units, unit in a course and standalone unit: https://github.com/code-dot-org/code-dot-org/blob/94199e6947ded834a9dd5eefa5a99f368433c953/dashboard/test/models/course_version_test.rb#L64-L73

I think this is sufficient coverage to feel confident with this change, but wanted to double check with you.